Add files via upload
Interactive interface to perform basic functions such as addition, subtraction, division, and
multiplication. It will need a display screen to display the user’s input and give the results. The grid
system in CSS is for the alignments of buttons on the calculator. The additional tools you need to
build a fully functional calculator include event Listeners, if-else statements, operators, loops, and so
on.  @https://github.com/AscentAcademy